+++ components/Preview.tsx
