+++ hooks/useEditor.tsx
@@ -1,12 +1,12 @@
 import dynamic from "next/dynamic";
-import "@/app/_styles/markdown-editor.css";
-import "@/app/_styles/markdown-preview.css";
+import "@/styles/markdown-editor.css";
+import "@/styles/markdown-preview.css";
 import { Commands } from "@uiw/react-markdown-editor/cjs/components/ToolBar";
 import { useCallback, useState } from "react";
 
 const MarkdownEditor = dynamic(
   () => import("@uiw/react-markdown-editor").then((mod) => mod.default),
-  { ssr: false }
+  { ssr: false },
 );
 
 const TOOLBARS: Commands[] = [
@@ -45,7 +45,7 @@ export default function useEditor(initContent: string = "Hello Markdown!") {
         toolbarsMode={["preview"]}
       />
     ),
-    [initContent]
+    [initContent],
   );
 
   return { content, Editor };
