+++ styles/markdown-editor.css
@@ -64,12 +64,14 @@
   padding-right: 5px;
 }
 [data-color-mode*="dark"] .md-editor {
-  --color-border-shadow: 0 0 0 1px rgba(255, 255, 255, 0.1),
-    0 0 0 rgb(255 255 255 / 0%), 0 1px 1px rgb(255 255 255 / 20%);
+  --color-border-shadow:
+    0 0 0 1px rgba(255, 255, 255, 0.1), 0 0 0 rgb(255 255 255 / 0%),
+    0 1px 1px rgb(255 255 255 / 20%);
 }
 [data-color-mode*="light"] .md-editor {
-  --color-border-shadow: 0 0 0 1px rgba(16, 22, 26, 0.1),
-    0 0 0 rgb(16 22 26 / 0%), 0 1px 1px rgb(16 22 26 / 20%);
+  --color-border-shadow:
+    0 0 0 1px rgba(16, 22, 26, 0.1), 0 0 0 rgb(16 22 26 / 0%),
+    0 1px 1px rgb(16 22 26 / 20%);
 }
 .md-editor {
   color: var(--color-fg-default);
